// src/enums/enums.ts
export enum Gender{
  MALE = 'MALE',
  FEMALE = 'FEMALE'
}

export enum PersonType {
  CHILD = 'CHILD',
  ADULT = 'ADULT'
}

export enum PatientWeight{
  MAX = 200,
  MIN = 1
}

export enum PatientHeight{
  MAX = 250,
  MIN = 10
}

export enum SubstanceAmount{
  ADULT_MAX = 220,
  CHILDREN_MAX = 120,
  MIN = 20
}

export enum TimeUnit{
  SECOND = 'SECOND',
  MINUTE = 'MINUTE',
  HOUR = 'HOUR',
  DAY = 'DAY'
}

export enum UserRole{
  USER = 'USER',
  ADMIN = 'ADMIN',
  SUPER_ADMIN = 'SUPER_ADMIN'
}